infra/requirements.txt
aws-cdk-lib >= 2.116.1, < 3